# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/itemku/product/update": {
      "post": {
        "description": "Updates the product identified by `id`. NOTE: this route carries @UserRoles but no @UseGuards(AuthGuard), so no token is checked today and the role restriction is not enforced.",
        "operationId": "ProductController_updateProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update an Itemku product price",
        "tags": [
          "Itemku Product"
        ]
      }
    },
...
    "/api/v1/product/create": {
      "post": {
        "description": "SuperAdmin only. Accepts an array; each item may reference one or more supplier products through `products`, which is how the sell price is composed.",
        "operationId": "ExternalProductController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateExtProduct"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create external products",
        "tags": [
          "External Product"
        ]
      }
    },
    "/api/v1/product/create-supplier": {
      "post": {
        "description": "SuperAdmin only. Every element is validated individually; a 400 with `errorCode` `4001` lists the offending array indexes and their validation messages.",
        "operationId": "ExternalProductController_createSupplierProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateSupplierProduct"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create supplier products",
        "tags": [
          "External Product"
        ]
      }
    },
    "/api/v1/product/update": {
      "post": {
        "description": "SuperAdmin only. The product is matched on `item_id`.",
        "operationId": "ExternalProductController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExtProduct"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update an external product",
        "tags": [
          "External Product"
        ]
      }
    },
    "/api/v1/product/get-list": {
      "post": {
        "description": "Paginated catalogue. The price returned depends on the caller role: admins see `admin_price`, SuperAdmins additionally see the supplier composition.",
        "operationId": "ExternalProductController_getAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GetExtProductListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List external products",
        "tags": [
          "External Product"
        ]
      }
    },
    "/api/v1/product/create-customer": {
      "post": {
        "description": "Admin only. Assigns per-customer selling prices for products the admin owns. `4007` when the target user is not a customer of the caller.",
        "operationId": "ExternalProductController_createCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CreateExtProductCustomer"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Set customer specific prices",
        "tags": [
          "External Product"
        ]
      }
    },
    "/api/v1/product/update-customer": {
      "post": {
        "description": "Admin only.",
        "operationId": "ExternalProductController_updateCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExtProductCustomer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update a customer specific price",
        "tags": [
          "External Product"
        ]
      }
    },
    "/api/v1/product/delete-customer": {
      "post": {
        "description": "Admin only.",
        "operationId": "ExternalProductController_deleteCustomer",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExtProductCustomer"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Remove a customer specific price",
        "tags": [
          "External Product"
        ]
      }
    },
    "/api/v1/product/update-supplier": {
      "post": {
        "description": "SuperAdmin only. Pulls the current Digiflazz price list and syncs it into supplier products. Takes no body.",
        "operationId": "ExternalProductController_updateDigi",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Refresh supplier prices from Digiflazz",
        "tags": [
          "External Product"
        ]
      }
    },
    "/api/v1/product/get-product": {
      "post": {
        "description": "SuperAdmin only. Returns the stored supplier catalogue. Takes no body.",
        "operationId": "ExternalProductController_getProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List supplier products",
        "tags": [
          "External Product"
        ]
      }
    },
...
  },
  "info": {
    "title": "TRX API Service",
    "description": "Transaction / top-up API service.\n\n## Authentication\n\nCall `POST /api/v1/auth/login` to obtain a JWT, then send it in the\n`authorization` header **without** the `Bearer ` prefix -- the guard verifies\nthe header value verbatim.\n\nMachine-to-machine endpoints (those under `/api/v1/tx/ext/...`) accept the\n`x-username` + `x-sign` header pair instead of a JWT. `x-sign` is the\nHMAC-SHA256 hex digest of the exact JSON request body signed with the\ncaller's API key (see `POST /api/v1/auth/api-key`).\n\nA few endpoints (`/api/v1/tx/request-web`, `/api/v1/external-topup/create-topup`,\n`/api/v1/external-topup/update-topup`) require BOTH a JWT and an `x-sign`\nheader over the body.\n\n## Response envelope\n\nMost endpoints answer with `{ message, data, errorCode }`. An `errorCode` of\n`\"0000\"` means success; any other code maps to an entry in\n`src/lib/errorCodes.ts` (e.g. `1000` wrong credential, `1002` insufficient\nbalance, `4012` signature not verified).",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization",
        "description": "Raw JWT returned by POST /api/v1/auth/login. Do NOT prefix it with \"Bearer \"."
      },
      "x-username": {
        "type": "apiKey",
        "in": "header",
        "name": "x-username",
        "description": "Username of the API consumer. Used together with x-sign."
      },
      "x-sign": {
        "type": "apiKey",
        "in": "header",
        "name": "x-sign",
        "description": "HMAC-SHA256 hex signature of the JSON request body."
      }
    },
    "schemas": {
...
      "ApiResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "success",
            "description": "Human readable result message"
          },
          "data": {
            "type": "object",
            "nullable": true,
            "description": "Endpoint specific payload. `null` when the call failed."
          },
          "errorCode": {
            "type": "string",
            "example": "0000",
            "description": "`0000` on success, otherwise a code from src/lib/errorCodes.ts"
          }
        },
        "required": [
          "message",
          "data"
        ]
      },
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 7,
            "description": "Id of the product price row to update"
          },
          "brand": {
            "type": "string",
            "example": "MOBILE LEGENDS"
          },
          "category": {
            "type": "string",
            "example": "Games"
          },
          "price": {
            "type": "number",
            "example": 25000,
            "description": "Selling price"
          },
          "buyer_sku_code": {
            "type": "string",
            "example": "ml86",
            "description": "Supplier SKU code"
          },
          "product_name": {
            "type": "string",
            "example": "Mobile Legends 86 Diamonds"
          },
          "actualPrice": {
            "type": "number",
            "example": 23000,
            "description": "Supplier price, used for margin checks"
          }
        },
        "required": [
          "id",
          "brand",
          "category",
          "price",
          "buyer_sku_code",
          "product_name",
          "actualPrice"
        ]
      },
...
      "ExtProductJunction": {
        "type": "object",
        "properties": {
          "item_id": {
            "type": "string",
            "example": "ITM-001",
            "description": "Parent external item id"
          },
          "product_id": {
            "type": "number",
            "example": 12,
            "description": "Id of the supplier product to draw from"
          },
          "qty": {
            "type": "number",
            "example": 2,
            "description": "How many units of the supplier product per order"
          }
        },
        "required": [
          "product_id",
          "qty"
        ]
      },
      "CreateExtProduct": {
        "type": "object",
        "properties": {
          "item_id": {
            "type": "string",
            "example": "ITM-001",
            "description": "External item id, unique per product"
          },
          "game_name": {
            "type": "string",
            "example": "Mobile Legends"
          },
          "item_name": {
            "type": "string",
            "example": "86 Diamonds"
          },
          "server_name": {
            "type": "string",
            "example": "Asia",
            "description": "Server / region label"
          },
          "group_name": {
            "type": "string",
            "example": "Diamond",
            "description": "Grouping label used in listings"
          },
          "stock": {
            "type": "number",
            "example": 100
          },
          "min_order": {
            "type": "number",
            "example": 1
          },
          "price": {
            "type": "number",
            "example": 25000,
            "description": "Price shown to customers"
          },
          "admin_price": {
            "type": "number",
            "example": 23000,
            "description": "Price charged to admins"
          },
          "products": {
            "description": "Supplier products this item is composed of.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ExtProductJunction"
            }
          }
        },
        "required": [
          "item_id",
          "game_name",
          "item_name",
          "server_name",
          "group_name",
          "stock",
          "min_order",
          "price",
          "admin_price",
          "products"
        ]
      },
      "CreateSupplierProduct": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Mobile Legends 86 Diamonds",
            "description": "Supplier product name"
          },
          "category": {
            "type": "string",
            "example": "Games"
          },
          "brand": {
            "type": "string",
            "example": "MOBILE LEGENDS"
          },
          "type": {
            "type": "string",
            "example": "Umum",
            "description": "Supplier product type"
          },
          "price": {
            "type": "number",
            "example": 23000,
            "description": "Current supplier price"
          },
          "code": {
            "type": "string",
            "example": "ml86",
            "description": "Supplier SKU code (buyer_sku_code)"
          },
          "status": {
            "type": "boolean",
            "example": true,
            "description": "Whether the product is sellable"
          },
          "unlimited_stock": {
            "type": "boolean",
            "example": false
          },
          "stock": {
            "type": "number",
            "example": 100,
            "description": "Ignored when unlimited_stock is true"
          },
          "multi": {
            "type": "boolean",
            "example": false,
            "description": "Supports multi transactions"
          },
          "start_cut_off": {
            "type": "string",
            "example": "23:00",
            "description": "Start of the daily cut off window (HH:mm)"
          },
          "end_cut_off": {
            "type": "string",
            "example": "01:00",
            "description": "End of the daily cut off window (HH:mm)"
          },
          "desc": {
            "type": "string",
            "example": "Instant process",
            "description": "Free-form supplier notes"
          },
          "actualPrice": {
            "type": "number",
            "example": 23000,
            "description": "Base price used for margin checks"
          }
        },
        "required": [
          "name",
          "category",
          "brand",
          "type",
          "price",
          "code",
          "status",
          "unlimited_stock",
          "stock",
          "multi",
          "start_cut_off",
          "end_cut_off",
          "desc",
          "actualPrice"
        ]
      },
      "ExtProductFilterDto": {
        "type": "object",
        "properties": {
          "item_id": {
            "type": "string",
            "example": "ITM-001"
          },
          "game_name": {
            "type": "string",
            "example": "Mobile Legends"
          },
          "item_name": {
            "type": "string",
            "example": "86 Diamonds"
          },
          "price": {
            "type": "number",
            "example": 25000,
            "description": "Exact price match. Compared against `admin_price` for Admin callers and `price` otherwise."
          }
        }
      },
      "GetExtProductListDto": {
        "type": "object",
        "properties": {
          "page": {
            "type": "number",
            "example": 1,
            "description": "Page number, 1 based"
          },
          "size": {
            "type": "number",
            "example": 10,
            "description": "Rows per page"
          },
          "filter": {
            "$ref": "#/components/schemas/ExtProductFilterDto"
          }
        }
      },
      "CreateExtProductCustomer": {
        "type": "object",
        "properties": {
          "item_id": {
            "type": "string",
            "example": "ITM-001",
            "description": "External item id of the product"
          },
          "cust_id": {
            "type": "number",
            "example": 42,
            "description": "Id of the customer this price applies to"
          },
          "price": {
            "type": "number",
            "example": 26000,
            "description": "Customer specific selling price"
          }
        },
        "required": [
          "item_id",
          "cust_id",
          "price"
        ]
      },
...
    }
  }
}